add list.misc
